requirements.txt
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
easyocr==1.7.2
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.16.1
Flask==3.1.0
Flask-MySQLdb==2.0.0
fonttools==4.55.3
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
gradio==5.9.0
gradio_client==1.5.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
imageio==2.36.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
mysqlclient==2.2.6
narwhals==1.18.4
networkx==3.4.2
ninja==1.11.1.3
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.12
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.2.3
pillow @ file:///D:/bld/pillow_1729065755620/work
protobuf==5.29.1
pyarrow==18.1.0
pyclipper==1.3.0.post6
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
pytesseract @ file:///home/conda/feedstock_root/build_artifacts/pytesseract_1697399219056/work
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruff==0.8.3
safehttpx==0.1.6
safetensors==0.4.5
scikit-image==0.25.0
scipy==1.14.1
semantic-version==2.10.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.3
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
tifffile==2024.12.12
timm==1.0.12
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
waitress==3.0.2
watchdog==6.0.0
websockets==14.1
Werkzeug==3.1.3
